dwback/requirements.txt
django-cors-headers
psycopg2
pillow
django-phonenumber-field[phonenumberslite]
django-random-image-from-folder
django-cryptography
